pyproject.toml
[tool.poetry]
name = "devops"
version = "0.1.0"
description = ""
authors = ["Emmanuel <emmanuel.davidson@thejitu.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
pymongo = "^4.2.0"
PyYAML = "^6.0"
inquirer = "^2.10.0"
uWSGI = "^2.0.20"
Flask-Cors = "^3.0.10"
aws-psycopg2 = "^1.3.8"
sqlparse = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
